Initial category data controller.
